Avoid including the same .cc file in srcs of multiple rules.
Ensure that the .cc files for variable ops kernels are only included in
the "srcs" of one rule ("variable_ops_kernels"), and change the other
rule that had them listed in srcs ("builtin_ops_kernels") to instead just
have a dependency on variable_ops_kernels. This is cleaner and avoids link
errors if you try to link both builtin_op_kernels and variable_op_kernels
into the same target.

PiperOrigin-RevId: 432179384